requirements.txt
# Core dependencies
gate-api>=4.20.0
aiohttp>=3.9.0
websockets>=12.0
asyncio>=3.4.3

# Database
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Telegram
aiogram>=3.4.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Monitoring
sentry-sdk>=1.40.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utils
orjson>=3.9.0
